refactor(wire): strongly-type BatchType + GcPauseType, tighten sink catches
Wire envelope now carries the BatchType enum directly instead of a raw byte,
matching the same pattern already used elsewhere. GcPauseType becomes a real
enum so call sites can no longer pass arbitrary bytes. UdpTelemetrySink's
SenderLoop and Dispose now catch only the exceptions that can actually escape
those code paths, surfacing send failures via SendErrors / LastSendError so
the collector daemon and tests can observe them. Slice helpers collapsed to a
single generic Slice<T> to drop duplicate copies.

Tests:
- RimObs.Library.Tests/UdpTelemetrySinkTests.cs covers null-owner guard, the
  initial error counter state, ring-buffer absorption without a receiver, the
  loopback drain path with a real UdpClient sink/receiver, and the
  send-to-unbound-port no-throw path.

+++ RimObs.Collector.Tests/EndToEndSmokeTests.cs
@@ -108,7 +108,7 @@ public sealed class EndToEndSmokeTests
             SchemaVersion = SchemaVersion.Current,
             Sequence = 1,
             OwnerId = "smoke",
-            BatchType = (byte)type,
+            BatchType = type,
             Payload = payload,
         };
         byte[] bytes = MessagePackSerializer.Serialize(envelope);
